$cdl fix, vol range fixes
$cdl was not being preserved across use_curl() calls because it was
being modified in a pipeline, and pipes use subshells.

Forgot to add largest vol range match to curl_oe_ls, so it wasn't
being added to the dl queue. Also, since the first volume in the
range is now handled within the dash matching block, we need to
initalize curl_oe_ls with it.