handle exact chapter ranges
I forgot to implement this. This code checks for a file chapter
range that matches an exact range the user requests. It is necessary
to add an extra bit of logic to ensure that we don't match a file
that has more chapters then the user requested. Unfortunately, this
still has some problems; A file with multiple ranges or a range
with extra chapters won't match the the latter chapters. This is
a good start, however.